third_party/blink/web_tests/external/wpt/editing/other/empty-elements-insertion-expected.txt

This is a testharness.js-based test.
[FAIL] Insert text into the unstyled inline element
  promise_test: Unhandled rejection with value: object "Error: element event-dispatch intercepted error"
[FAIL] Insert text into the unstyled inline element with the styled ::before pseudoelement
  assert_greater_than: The text should be inserted into the <strong> element expected a number greater than 0 but got 0
[FAIL] Insert text into the unstyled inline element with the styled ::after pseudoelement
  assert_greater_than: The text should be inserted into the <strong> element expected a number greater than 0 but got 0
[FAIL] Insert text into the unstyled inline element with the styled ::before and ::after pseudoelements
  assert_greater_than: The text should be inserted into the <strong> element expected a number greater than 0 but got 0
Harness: the test ran to completion.
